Added %u

+++ src/c/interrupts/handler.c
@@ -47,7 +47,7 @@ cpu_state_t *int_handler(cpu_state_t *cpu)
 				puts("Stack Fault");
 				break;
 			case 0x0d:
-				printf("General Protection Fault\nError Code: %u\n",cpu->error);
+				printf("General Protection Fault\nError Code: %u\n",(unsigned int)cpu->error);
 				while(1);
 				break;
 			case 0x0e:
